hackathon_challenge/scripts/run_simulation.py
#!/usr/bin/env python3
"""Run QKD protocol simulation.

Reference: implementation_plan.md §3.3
"""

import yaml

# TODO: Import after SquidASM is available
# from squidasm.run.stack.run import run
# from squidasm.util.log import LogManager

# TODO: Import after implementation
# from hackathon_challenge.core.protocol import AliceProgram, BobProgram


def main() -> None:
    """Execute QKD simulation and analyze results."""
    # Load configuration
    with open("config.yaml", "r") as f:
        config = yaml.safe_load(f)

    print("QKD Simulation")
    print("=" * 60)
    print(f"Configuration loaded: {config['simulation']['num_times']} runs")
    print()

    # TODO: Implement simulation execution
    # 1. Setup logging
    # 2. Create program instances
    # 3. Run simulation
    # 4. Analyze results

    print("TODO: Implementation in progress")
    print("Reference: implementation_plan.md §Phase 5")


if __name__ == "__main__":
    main()
