hackathon_challenge/pyproject.toml
[project]
name = "qkd-extension"
version = "0.1.0"
description = "Extended QKD implementation with Cascade reconciliation and Toeplitz privacy amplification"
requires-python = ">=3.9"
authors = [
    {name = "QIA Hackathon 2025 Team"}
]

dependencies = [
    "squidasm>=0.11.0",
    "netqasm>=0.14.0",
    "numpy>=1.23.0",
    "scipy>=1.9.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "black>=23.0",
    "flake8>=6.0",
]

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["hackathon_challenge"]
omit = ["*/tests/*", "*/scripts/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
